Move student dashboard view to the right file
